[REFACTOR] Change the balance primitive type for big decimal, because this field about monetary value

+++ src/main/java/com/tech/padawan/financialmanager/transaction/dto/UpdateTransactionDTO.java
@@ -4,10 +4,12 @@ import com.tech.padawan.financialmanager.transaction.model.TransactionCategory;
 import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotNull;
 
+import java.math.BigDecimal;
+
 public record UpdateTransactionDTO(
         @NotNull(message = "Value is required")
         @Min(value = 0, message = "Value most be a positive number")
-        double value,
+        BigDecimal value,
         String description,
         @NotNull(message = "Category is required")
         Long categoryId
